+++ src/entities/category/index.ts
@@ -1,0 +1,31 @@
+export { categoryApi } from './model/api'
+export { categoryRepository } from './local/repository'
+export {
+  enqueueCreateCategory,
+  enqueueUpdateCategory,
+  enqueueDeleteCategory,
+  enqueueReorderCategories,
+  isTempId,
+} from './local/outbox-helpers'
+export {
+  fetchCategories,
+  createCategory,
+  updateCategory,
+  reorderCategories,
+  deleteCategory,
+  setCategories,
+  addCategory,
+  removeCategory,
+  default as categoryReducer,
+} from './model/category.slice'
+export {
+  selectCategories,
+  selectCategoriesStatus,
+  selectCategoryById,
+} from './model/selectors'
+export type {
+  CategoryDto,
+  CategoryFormValues,
+  CreateCategoryDto,
+  UpdateCategoryDto,
+} from './model/types'