+++ src/features/category/index.ts
@@ -1,4 +1,3 @@
-export { categoryApi } from './api'
 export { CategoryForm } from './components/CategoryForm'
 export { CategoryCard } from './components/CategoryCard'
 export { CategoryPicker } from './components/CategoryPicker'