+++ src/store/slices/app/index.ts
@@ -1,6 +1,6 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit'
 
-import { fetchCategories } from '@/store/slices/category'
+import { fetchCategories } from '@/entities/category'
 import { fetchTags } from '@/entities/tag'
 import { fetchAccounts } from '@/store/slices/account'
 import { fetchSettings } from '@/entities/settings'