+++ src/store/store.ts
@@ -2,11 +2,11 @@ import { configureStore } from '@reduxjs/toolkit'
 
 import { userReducer } from '@/entities/user'
 import { tagReducer } from '@/entities/tag'
+import { categoryReducer } from '@/entities/category'
 
 import transactionTypeReducer from './slices/transaction-type'
 import timeRangeReducer from './slices/time-range'
 import appReducer from './slices/app'
-import categoryReducer from './slices/category'
 import accountReducer from './slices/account'
 import settingsReducer from './slices/settings'
 import transactionReducer from './slices/transaction'